Converts keycloakdeployment to a dependent resource
Closes #22225
